requirements.txt
Flask
flask-cors
Pillow
numpy
torch
realesrgan
basicsr
gunicorn
